+++ packages/next-adapter/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @crossed/next-adapter
 
+## 0.6.3
+
+### Patch Changes
+
+- Updated dependencies
+  - @crossed/webpack@2.0.3
+
 ## 0.6.2
 
 ### Patch Changes